Add files via upload
The original dataset containing the Primitive Lottery and the Bonoloto winning lottery numbers from 1988 to 2006 could not be located. The New Zealand Poweball lottery numbers were chosen as a subsequent dataset to base the visualization on. The original Poweball numbers can be found here: https://assets.mylotto.co.nz/assets/uploads/d89aa3aa-306a-11e8-9184-a0842176f481.xlsx

This dataset, was preprocessed to subset for all (6) winning numbers; that is, excluding all bonus numbers and the powerball number from 1988-2006.  The file contains a total of 6156 observations with 2 variables. Year is a factor with 19 levels. Numbers is a factor with 40 levels.